requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==3.1.2
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
openai==0.28.1
packaging==25.0
propcache==0.4.1
PyMuPDF==1.26.5
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
tqdm==4.67.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
